Moved RangeSelector.
