+++ ts/masters/modules/gantt.src.ts
@@ -15,13 +15,15 @@ import XRangeSeries from '../../Series/XRange/XRangeSeries.js';
 import '../../Series/Gantt/GanttSeries.js';
 import GanttChart from '../../Core/Chart/GanttChart.js';
 import Scrollbar from '../../Core/Scrollbar.js';
-import '../../Extensions/RangeSelector.js';
+import RangeSelector from '../../Stock/RangeSelector/RangeSelector.js';
 import '../../Core/Navigator.js';
 const G: AnyRecord = Highcharts;
 // Classes
-G.Scrollbar = Scrollbar;
 G.GanttChart = GanttChart;
 G.ganttChart = GanttChart.ganttChart;
+G.RangeSelector = RangeSelector;
+G.Scrollbar = Scrollbar;
 // Compositions
 Scrollbar.compose(G.Axis);
+RangeSelector.compose(G.Axis, G.Chart);
 XRangeSeries.compose(G.Axis);
